fix: broken annotated commit detection
Annotated commits have their own object hash in addition to the commit
hash. gitRepo.Tag() returns the hash of the tag instead of the commit.
If this resulting hash is passed to CommitsBetween, as is done in
`chglog add`, the hash will never be found. Update GitHashFotTag to
return the commit hash if the tag is an annotated tag.

This fixes an issue where `chglog add` adds the entire git history when
annotated tags are used.